+++ bot/handlers/personal_acount.py
@@ -1,32 +1,22 @@
 from aiogram import F, Router
 from aiogram.fsm.context import FSMContext
-from aiogram.fsm.state import State, StatesGroup
 from aiogram.types import CallbackQuery, Message
 
-from bot.handlers.launch_bot import ADMIN_USER_ID
-from bot.utils.read_text import load_text_form_file
 from bot.database.database import (
     get_user_data,  # Импорт функции получения пользователя из базы
     update_user_data,  # Импорт функции изменения данных пользователя в базе
 )
-from bot.keyboards.keyboards import (
-    create_data_change_buttons,
-    generate_authorized_user_options_keyboard,
-    generate_keyboard_personal_account,
-    generate_user_options_keyboard,
-    generate_admin_button,
-)
+from bot.handlers.launch_bot import ADMIN_USER_ID
+from bot.keyboards.keyboards import (create_data_change_buttons, generate_authorized_user_options_keyboard,
+                                     generate_keyboard_personal_account,
+                                     generate_user_options_keyboard,
+                                     generate_admin_button)
+from bot.states.states import ChangeData
+from bot.utils.read_text import load_text_form_file
 
 router_personal_acount = (
     Router()
 )  # Создание маршрутизатора для обработки команд и сообщений.
-
-
-class ChangeData(StatesGroup):
-    name = State()  # Состояние ввода для изменения имени.
-    height = State()  # Состояние ввода для изменения роста.
-    weight = State()  # Состояние ввода для изменения веса.
-    training_experience = State()  # Состояние ввода для изменения опыта тренировок.
 
 
 # Обработчик состояния просмотра личного кабинета
@@ -67,9 +57,7 @@ async def back_to_personal_account(callback_query: CallbackQuery) -> None:
 
 # Обработчик состояния изменения имя профиля
 @router_personal_acount.callback_query(F.data == "update_name")
-async def update_user_data_name(
-        callback_query: CallbackQuery, state: FSMContext
-) -> None:
+async def update_user_data_name(callback_query: CallbackQuery, state: FSMContext) -> None:
     user_id = callback_query.from_user.id
     data_user = get_user_data(user_id)
     if data_user:
@@ -97,9 +85,7 @@ async def update_name(message: Message, state: FSMContext) -> None:
 
 # Обработчик состояния изменения рост профиля
 @router_personal_acount.callback_query(F.data == "update_height")
-async def update_user_data_height(
-        callback_query: CallbackQuery, state: FSMContext
-) -> None:
+async def update_user_data_height(callback_query: CallbackQuery, state: FSMContext) -> None:
     user_id = callback_query.from_user.id
     data_user = get_user_data(user_id)
     if data_user:
@@ -127,9 +113,7 @@ async def update_height(message: Message, state: FSMContext) -> None:
 
 # Обработчик состояния изменения вес профиля
 @router_personal_acount.callback_query(F.data == "update_weight")
-async def update_user_data_weight(
-        callback_query: CallbackQuery, state: FSMContext
-) -> None:
+async def update_user_data_weight(callback_query: CallbackQuery, state: FSMContext) -> None:
     user_id = callback_query.from_user.id
     data_user = get_user_data(user_id)
     if data_user:
@@ -157,9 +141,7 @@ async def update_weight(message: Message, state: FSMContext) -> None:
 
 # Обработчик состояния изменения опыт тренировок профиля
 @router_personal_acount.callback_query(F.data == "update_training_experience")
-async def update_user_data_training_experience(
-        callback_query: CallbackQuery, state: FSMContext
-) -> None:
+async def update_user_data_training_experience(callback_query: CallbackQuery, state: FSMContext) -> None:
     user_id = callback_query.from_user.id
     data_user = get_user_data(user_id)
     if data_user:
